// projects/dashboard/src/app/interfaces/oanda/oanda.ts
export interface Order {
// ...
  type: OrderType;
// ...
}

/** Get a list of Orders for an Account */
export interface Orders {
  /**The list of Order detail objects */
  orders: Order[];

  /**
   * The ID of the most recent Transaction created for the Account.
   */
  lastTransactionID: string;
}
// ...
type GuaranteedStopLossOrderMode = 'DISABLED' | 'ALLOWED' | 'REQUIRED';

export interface AccountSummaryResponse {
  /**
   * The summary of the requested Account.
   */
  account: AccountSummary;

  /**
   * The ID of the most recent Transaction created for the Account.
   */
  lastTransactionID: string;
}

/**
 * The full details of a client's Account. This includes full open Trade, open
 * Position and pending Order representation.
 */
export interface Account {
  /**
   * The Account's identifier
   */
  id: string;

  /**
   * Client-assigned alias for the Account. Only provided if the Account has
   * an alias set
   */
  alias: string;

  /**
   * The home currency of the Account
   */
  currency: string;

  /**
   * ID of the user that created the Account.
   */
  createdByUserID: number;

  /**
   * The date/time when the Account was created.
   */
  createdTime: string;

  /**
   * The current guaranteed Stop Loss Order settings of the Account. This
   * field will only be present if the guaranteedStopLossOrderMode is not
   * ‘DISABLED'.
   */
  guaranteedStopLossOrderParameters: GuaranteedStopLossOrderParameters;

  /**
   * The current guaranteed Stop Loss Order mode of the Account.
   */
  guaranteedStopLossOrderMode: GuaranteedStopLossOrderMode;

  /**
   * The date/time that the Account's resettablePL was last reset.
   */
  resettablePLTime: string;

  /**
   * Client-provided margin rate override for the Account. The effective
   * margin rate of the Account is the lesser of this value and the OANDA
   * margin rate for the Account's division. This value is only provided if a
   * margin rate override exists for the Account.
   */
  marginRate: string;

  /**
   * The number of Trades currently open in the Account.
   */
  openTradeCount: number;

  /**
   * The number of Positions currently open in the Account.
   */
  openPositionCount: number;

  /**
   * The number of Orders currently pending in the Account.
   */
  pendingOrderCount: number;

  /**
   * Flag indicating that the Account has hedging enabled.
   */
  hedgingEnabled: boolean;

  /**
   * The total unrealized profit/loss for all Trades currently open in the
   * Account.
   */
  unrealizedPL: string;

  /**
   * The net asset value of the Account. Equal to Account balance +
   * unrealizedPL.
   */
  NAV: string;

  /**
   * Margin currently used for the Account.
   */
  marginUsed: string;

  /**
   * Margin available for Account currency.
   */
  marginAvailable: string;

  /**
   * The value of the Account's open positions represented in the Account's
   * home currency.
   */
  positionValue: string;

  /**
   * The Account's margin closeout unrealized PL.
   */
  marginCloseoutUnrealizedPL: string;

  /**
   * The Account's margin closeout NAV.
   */
  marginCloseoutNAV: string;

  /**
   * The Account's margin closeout margin used.
   */
  marginCloseoutMarginUsed: string;

  /**
   * The Account's margin closeout percentage. When this value is 1.0 or above
   * the Account is in a margin closeout situation.
   */
  marginCloseoutPercent: string;

  /**
   * The value of the Account's open positions as used for margin closeout
   * calculations represented in the Account's home currency.
   */
  marginCloseoutPositionValue: string;

  /**
   * The current WithdrawalLimit for the account which will be zero or a
   * positive value indicating how much can be withdrawn from the account.
   */
  withdrawalLimit: string;

  /**
   * The Account's margin call margin used.
   */
  marginCallMarginUsed: string;

  /**
   * The Account's margin call percentage. When this value is 1.0 or above the
   * Account is in a margin call situation.
   */
  marginCallPercent: string;

  /**
   * The current balance of the account.
   */
  balance: string;

  /**
   * The total profit/loss realized over the lifetime of the Account.
   */
  pl: string;

  /**
   * The total realized profit/loss for the account since it was last reset by
   * the client.
   */
  resettablePL: string;

  /**
   * The total amount of financing paid/collected over the lifetime of the
   * account.
   */
  financing: string;

  /**
   * The total amount of commission paid over the lifetime of the Account.
   */
  commission: string;

  /**
   * The total amount of dividend adjustment paid over the lifetime of the
   * Account in the Account's home currency.
   */
  dividendAdjustment: string;

  /**
   * The total amount of fees charged over the lifetime of the Account for the
   * execution of guaranteed Stop Loss Orders.
   */
  guaranteedExecutionFees: string;

  /**
   * The date/time when the Account entered a margin call state. Only provided
   * if the Account is in a margin call.
   */
  marginCallEnterTime: string;

  /**
   * The number of times that the Account's current margin call was extended.
   */
  marginCallExtensionCount: number;

  /**
   * The date/time of the Account's last margin call extension.
   */
  lastMarginCallExtensionTime: string;

  /**
   * The ID of the last Transaction created for the Account.
   */
  lastTransactionID: string;

  /**
   * The details of the Trades currently open in the Account.
   */
  trades: TradeSummary[];

  /**
   * The details all Account Positions.
   */
  positions: Position[];

  /**
   * The details of the Orders currently pending in the Account.
   */
  orders: Order[];
}
// ...
export interface TransactionsResponse {
  /**
   * The starting time provided in the request.
   */
  from: string;

  /**
   * The ending time provided in the request.
   */
  to: string;

  /**
   * The pageSize provided in the request
   */
  pageSize: number;

  /**
   * The Transaction-type filter provided in the request
   */
  type: TransactionFilter[];

  /**
   * The number of Transactions that are contained in the pages returned
   */
  count: number;

  /**
   * The list of URLs that represent idrange queries providing the data for
   * each page in the query results
   */
  pages: string[];

  /**
   * The ID of the most recent Transaction created for the Account
   */
  lastTransactionID: string;
}

/**
 * Transaction Response
 */
export interface TransactionResponse {
  /**
   * The details of the Transaction requested
   */
  transaction: Transaction;

  /**
   * The ID of the most recent Transaction created for the Account
   */
  lastTransactionID: string;
}

export interface Transaction {
  /**
   * The Transaction's Identifier.
   */
  id: string;

  /**
   * The Orders Instrument.
   */
  instrument: string;

  /**
   * The date/time when the Transaction was created.
   */
  time: string;

  /**
   * The ID of the user that initiated the creation of the Transaction.
   */
  userID: number;

  /**
   * The ID of the Account the Transaction was created for.
   */
  accountID: string;

  /**
   * The ID of the “batch” that the Transaction belongs to. Transactions in
   * the same batch are applied to the Account simultaneously.
   */
  batchID: string;

  /**
   * The Request ID of the request which generated the transaction.
   */
  requestID?: string;

  /**
   * The Type of the Transaction.
   */
  type: TransactionType | string;
}
// ...
export enum TransactionType {
  CREATE = 'Account Create Transaction',
  CLOSE = 'Account Close Transaction',
  REOPEN = 'Account Reopen Transaction',
  CLIENT_CONFIGURE = 'Client Configuration Transaction',
  CLIENT_CONFIGURE_REJECT = 'Client Configuration Reject Transaction',
  TRANSFER_FUNDS = 'Transfer Funds Transaction',
  TRANSFER_FUNDS_REJECT = 'Transfer Funds Reject Transaction',
  MARKET_ORDER = 'Market Order Transaction',
  MARKET_ORDER_REJECT = 'Market Order Reject Transaction',
  FIXED_PRICE_ORDER = 'Fixed Price Order Transaction',
  LIMIT_ORDER = 'Limit Order Transaction',
  LIMIT_ORDER_REJECT = 'Limit Order Reject Transaction',
  STOP_ORDER = 'Stop Order Transaction',
  STOP_ORDER_REJECT = 'Stop Order Reject Transaction',
  MARKET_IF_TOUCHED_ORDER = 'Market if Touched Order Transaction',
  MARKET_IF_TOUCHED_ORDER_REJECT = 'Market if Touched Order Reject Transaction',
  TAKE_PROFIT_ORDER = 'Take Profit Order Transaction',
  TAKE_PROFIT_ORDER_REJECT = 'Take Profit Order Reject Transaction',
  STOP_LOSS_ORDER = 'Stop Loss Order Transaction',
  STOP_LOSS_ORDER_REJECT = 'Stop Loss Order Reject Transaction',
  GUARANTEED_STOP_LOSS_ORDER = 'Guaranteed Stop Loss Order Transaction',
  GUARANTEED_STOP_LOSS_ORDER_REJECT = 'Guaranteed Stop Loss Order Reject Transaction',
  TRAILING_STOP_LOSS_ORDER = 'Trailing Stop Loss Order Transaction',
  TRAILING_STOP_LOSS_ORDER_REJECT = 'Trailing Stop Loss Order Reject Transaction',
  ORDER_FILL = 'Order Fill Transaction',
  ORDER_CANCEL = 'Order Cancel Transaction',
  ORDER_CANCEL_REJECT = 'Order Cancel Reject Transaction',
  ORDER_CLIENT_EXTENSIONS_MODIFY = 'Order Client Extensions Modify Transaction',
  ORDER_CLIENT_EXTENSIONS_MODIFY_REJECT = 'Order Client Extensions Modify Reject Transaction',
  TRADE_CLIENT_EXTENSIONS_MODIFY = 'Trade Client Extensions Modify Transaction',
  TRADE_CLIENT_EXTENSIONS_MODIFY_REJECT = 'Trade Client Extensions Modify Reject Transaction',
  MARGIN_CALL_ENTER = 'Margin Call Enter Transaction',
  MARGIN_CALL_EXTEND = 'Margin Call Extend Transaction',
  MARGIN_CALL_EXIT = 'Margin Call Exit Transaction',
  DELAYED_TRADE_CLOSURE = 'Delayed Trade Closure Transaction',
  DAILY_FINANCING = 'Daily Financing Transaction',
  DIVIDEND_ADJUSTMENT = 'Dividend Adjustment Transaction',
  RESET_RESETTABLE_PL = 'Reset Resettable PL Transaction',
}
// ...
export enum TransactionFilter {
  ORDER = 'Order-related Transactions. These are the Transactions that create, cancel, fill or trigger Orders',
  FUNDING = 'Funding-related Transactions',
  ADMIN = 'Administrative Transactions',
  CREATE = 'Account Create Transaction',
  CLOSE = 'Account Close Transaction',
  REOPEN = 'Account Reopen Transaction',
  CLIENT_CONFIGURE = 'Client Configuration Transaction',
  CLIENT_CONFIGURE_REJECT = 'Client Configuration Reject Transaction',
  TRANSFER_FUNDS = 'Transfer Funds Transaction',
  TRANSFER_FUNDS_REJECT = 'Transfer Funds Reject Transaction',
  MARKET_ORDER = 'Market Order Transaction',
  MARKET_ORDER_REJECT = 'Market Order Reject Transaction',
  LIMIT_ORDER = 'Limit Order Transaction',
  LIMIT_ORDER_REJECT = 'Limit Order Reject Transaction',
  STOP_ORDER = 'Stop Order Transaction',
  STOP_ORDER_REJECT = 'Stop Order Reject Transaction',
  MARKET_IF_TOUCHED_ORDER = 'Market if Touched Order Transaction',
  MARKET_IF_TOUCHED_ORDER_REJECT = 'Market if Touched Order Reject Transaction',
  TAKE_PROFIT_ORDER = 'Take Profit Order Transaction',
  TAKE_PROFIT_ORDER_REJECT = 'Take Profit Order Reject Transaction',
  STOP_LOSS_ORDER = 'Stop Loss Order Transaction',
  STOP_LOSS_ORDER_REJECT = 'Stop Loss Order Reject Transaction',
  GUARANTEED_STOP_LOSS_ORDER = 'Guaranteed Stop Loss Order Transaction',
  GUARANTEED_STOP_LOSS_ORDER_REJECT = 'Guaranteed Stop Loss Order Reject Transaction',
  TRAILING_STOP_LOSS_ORDER = 'Trailing Stop Loss Order Transaction',
  TRAILING_STOP_LOSS_ORDER_REJECT = 'Trailing Stop Loss Order Reject Transaction',
  ONE_CANCELS_ALL_ORDER = 'One Cancels All Order Transaction',
  ONE_CANCELS_ALL_ORDER_REJECT = 'One Cancels All Order Reject Transaction',
  ONE_CANCELS_ALL_ORDER_TRIGGERED = 'One Cancels All Order Trigger Transaction',
  ORDER_FILL = 'Order Fill Transaction',
  ORDER_CANCEL = 'Order Cancel Transaction',
  ORDER_CANCEL_REJECT = 'Order Cancel Reject Transaction',
  ORDER_CLIENT_EXTENSIONS_MODIFY = 'Order Client Extensions Modify Transaction',
  ORDER_CLIENT_EXTENSIONS_MODIFY_REJECT = 'Order Client Extensions Modify Reject Transaction',
  TRADE_CLIENT_EXTENSIONS_MODIFY = 'Trade Client Extensions Modify Transaction',
  TRADE_CLIENT_EXTENSIONS_MODIFY_REJECT = 'Trade Client Extensions Modify Reject Transaction',
  MARGIN_CALL_ENTER = 'Margin Call Enter Transaction',
  MARGIN_CALL_EXTEND = 'Margin Call Extend Transaction',
  MARGIN_CALL_EXIT = 'Margin Call Exit Transaction',
  DELAYED_TRADE_CLOSURE = 'Delayed Trade Closure Transaction',
  DAILY_FINANCING = 'Daily Financing Transaction',
  RESET_RESETTABLE_PL = 'Reset Resettable PL Transaction',
}

export const transactionFilterList: {
  key: string;
  value: string;
}[] = Object.entries(TransactionFilter).map(([key, value]) => ({ key, value }));

export interface TransactionFilterList {
  key: TransactionFilter | string;
  value: string;
}